refactor(mm-next): move logGqlError
